change RPN to double and ex02 of Module09

+++ Module_09/ex01/RPN.hpp
@@ -14,7 +14,7 @@ private:
 	RPN &operator=(const RPN &rpn);
 	~RPN();
 public:
-	static int calculateRPN(std::string expression);
+	static double calculateRPN(std::string expression);
 };
